Move RenderTemplate to testutils package
This helper is only used in tests.
