+++ _about/CONTRIBUTING.md
@@ -206,7 +206,7 @@ to check under `../internal/client/*.go`.
 Test fixtures are used to create resources in the test environment. They are typically used to create resources
 that are used in the test, such as a workspace, account, or deployment.
 
-The `internal/provider/helpers` package contains a helper function named `RenderTemplate` that can be used to
+The `internal/provider/testutils` package contains a helper function named `RenderTemplate` that can be used to
 create test fixtures that contain multiple resources. This function is especially useful for creating test fixtures that
 contain multiple resources, making it easier to visually understand where each variable is inserted when compared to
 using `fmt.Sprintf`.
