+++ internal/provider/datasources/automation_test.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/statecheck"
-	"github.com/prefecthq/terraform-provider-prefect/internal/provider/helpers"
 	"github.com/prefecthq/terraform-provider-prefect/internal/testutils"
 )
 
@@ -69,7 +68,7 @@ data "prefect_automation" "{{ .AutomationResourceName }}" {
 }
 `
 
-	return helpers.RenderTemplate(tmpl, cfg)
+	return testutils.RenderTemplate(tmpl, cfg)
 }
 
 func fixtureAccAutomationResourceMetricTrigger(cfg automationFixtureConfig) string {
@@ -118,7 +117,7 @@ data "prefect_automation" "{{ .AutomationResourceName }}" {
 }
 `
 
-	return helpers.RenderTemplate(tmpl, cfg)
+	return testutils.RenderTemplate(tmpl, cfg)
 }
 
 func fixtureAccAutomationResourceCompoundTrigger(cfg automationFixtureConfig) string {
@@ -190,7 +189,7 @@ data "prefect_automation" "{{ .AutomationResourceName }}" {
 }
 `
 
-	return helpers.RenderTemplate(tmpl, cfg)
+	return testutils.RenderTemplate(tmpl, cfg)
 }
 
 func fixtureAccAutomationResourceSequenceTrigger(cfg automationFixtureConfig) string {
@@ -269,7 +268,7 @@ data "prefect_automation" "{{ .AutomationResourceName }}" {
 }
 `
 
-	return helpers.RenderTemplate(tmpl, cfg)
+	return testutils.RenderTemplate(tmpl, cfg)
 }
 
 //nolint:paralleltest // we use the resource.ParallelTest helper instead
