+++ internal/testutils/template.go
@@ -1,4 +1,4 @@
-package helpers
+package testutils
 
 import (
 	"bytes"
